+++ src/dyadic.jl
@@ -64,18 +64,20 @@ function _inv(d::DyadicReal, ::RoundingMode{:Nearest}; precision = DEFAULT_PRECI
 end
 
 function Base.inv(
-        d::DyadicReal, r::RoundingMode = RoundNearest; precision = DEFAULT_PRECISION)
-    0 ∈ d && throw(DivideError())
-    if iszero(abs(d.m) & (abs(d.m) - 1))
-        DyadicReal(big(1) << d.e, ilog2(d.m))
-    else
-        _inv(d, r; precision)
-    end
+        d::DyadicReal; precision = DEFAULT_PRECISION)
+    iszero(d) && throw(DivideError())
+    RationalCauchyCut(big(1) << d.e, d.m)
+    # if iszero(abs(d.m) & (abs(d.m) - 1))
+    #     DyadicReal(big(1) << d.e, ilog2(d.m))
+    # else
+    #     _inv(d, r; precision)
+    # end
 end
 
 function Base.:/(
-        d1::DyadicReal, d2::DyadicReal, r = RoundNearest; precision = DEFAULT_PRECISION)
-    d1 * inv(d2, r; precision)
+        d1::DyadicReal, d2::DyadicReal; precision = DEFAULT_PRECISION)
+    iszero(d2) && throw(Divideerror())
+    RationalCauchyCut((d1.m << d2.e) // (d2.m << d1.e))
 end
 
 #################