+++ docs/src/changelog.md
@@ -10,7 +10,7 @@ EditURL = "https://github.com/lucaferranti/DedekindCutArithmetic.jl/blob/main/CH
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [v0.1.1](https://github.com/lucaferranti/DedekindCutArithmetic.jl/releases/tag/v0.1.1) -- 2025-04-13
 
 ### Added